Cast stack sizes to size_t to fix errors on some compilers
The size of a stack trace is compiler defined, so depending on how the
compiler defines it, it might not be the same as the type of
m_max_trace. On my setup, g++ 14.2.1 x86_64-linux, m_stack_trace is a
long unsigned int (size_t) and stack.size() is a short unsigned int. std::min
requires that both arguments be the same type to properly instantiate
the template, so this produces a compiler error.

My fix is to cast the stack size to size_t. On my machine at least,
this is strictly bigger than it was, so this doesn't create any new
danger of overflow.

+++ include/core/log.hpp
@@ -62,7 +62,8 @@ namespace Toolbox::Log {
 
         void trace(const std::stacktrace &stack) {
 #ifdef TOOLBOX_DEBUG
-            for (size_t i = 0; i < std::min(m_max_trace, stack.size()); ++i) {
+            for (size_t i = 0; i < std::min(m_max_trace,
+                                            (size_t)stack.size()); ++i) {
                 debugLog(
                     std::format("{} at line {}", stack[i].source_file(), stack[i].source_line()));
             }
